Trabalho Mamografia/requirements.txt
## The following requirements were added by pip freeze:
cycler==0.11.0
fonttools==4.33.3
imageio==2.19.1
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.2
networkx==2.8
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.9
PySimpleGUI==4.60.0
python-dateutil==2.8.2
PyWavelets==1.3.0
scikit-image==0.19.2
scikit-learn==1.1.0
scipy==1.8.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tifffile==2022.5.4
